+++ RumahSakit/src/rumahsakit/DataDokter.java
@@ -21,10 +21,10 @@ ResultSet RsProduk=null;
      */
     public DataDokter() {
         initComponents();
-        tampilData();
+        showData();
     }
 
-        private void tampilData(){
+        private void showData(){
         try{
             Object[] judul_kolom = {"id_dokter", "nik", "nama", "alamat", "telepon", "tglahir", "goldar", "gender", "spesialisasi"};
             tabModel=new DefaultTableModel(null,judul_kolom);
@@ -54,8 +54,9 @@ ResultSet RsProduk=null;
         }
     }
    
-//menampilkan data ke form saat data pada tabel di klik 
-    void tblKeForm(){
+    //show data to form when click data on table
+    //menampilkan data ke form saat data pada tabel di klik 
+    void tableToForm(){
         jTextFieldNama.setText(tabModel.getValueAt(TabelDokter.getSelectedRow(), 2)+ "");
         jTextFieldAlamat.setText(tabModel.getValueAt(TabelDokter.getSelectedRow(), 3)+ "");
         jTextFieldTelepon.setText(tabModel.getValueAt(TabelDokter.getSelectedRow(), 4)+ "");
@@ -70,8 +71,9 @@ ResultSet RsProduk=null;
         buttonSave.setEnabled(false);
     }
     
-   //membersihkan form
-    private void BersihData(){
+    //clear the form
+    //membersihkan isian form 
+    private void clearData(){
         
         jTextFieldNIK.setText("");
         jTextFieldNama.setText(""); 
@@ -84,7 +86,8 @@ ResultSet RsProduk=null;
         
     } 
     
-   //disable form
+    //disable to fill in form
+    //tidak membolehkan untuk mengisi form
     private void SetEditOff(){
         jTextFieldNIK.setEnabled(false);
         jTextFieldNama.setEnabled(false); 
@@ -96,7 +99,8 @@ ResultSet RsProduk=null;
         jTextFieldSpesialisasi.setEnabled(false);
     }
     
-    //enable form
+    //enable to fill in form
+    //membolehkan untuk mengisi form
     private void seteditOn(){
         jTextFieldNIK.setEnabled(true);
         jTextFieldNama.setEnabled(true); 
@@ -394,16 +398,20 @@ ResultSet RsProduk=null;
     private void jTextFieldNIKActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jTextFieldNIKActionPerformed
         // TODO add your handling code here:
     }//GEN-LAST:event_jTextFieldNIKActionPerformed
-
+    
+    //process for making new data
+    //proses untuk membuat data data
     private void buttonNewActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonNewActionPerformed
         // TODO add your handling code here:
-        BersihData();
+        clearData();
         buttonSave.setEnabled(true);
         buttonUpdate.setEnabled(false);
         buttonDelete.setEnabled(false);
         seteditOn();
     }//GEN-LAST:event_buttonNewActionPerformed
 
+    //process for deleting data
+    //proses untuk menghapus data
     private void buttonDeleteActionPerformed(java.awt.event.ActionEvent evt) {                                             
         // TODO add your handling code here:
         String nama=jTextFieldNama.getText();
@@ -417,8 +425,8 @@ ResultSet RsProduk=null;
             Connection conn=(Connection)koneksi.koneksiDB();
             Statement stt=conn.createStatement();
             stt.executeUpdate("DELETE FROM dokter WHERE nama='"+nama+"'");
-            BersihData();
-            tampilData();
+            clearData();
+            showData();
             SetEditOff();
             JOptionPane.showMessageDialog(this,"Data berhasil di hapus","Success",JOptionPane.INFORMATION_MESSAGE);
         } catch(SQLException e){
@@ -427,6 +435,8 @@ ResultSet RsProduk=null;
         }
     }
     
+    //process for updating data
+    //proses untuk memperbaharui data
     private void buttonUpdateActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonUpdateActionPerformed
         // TODO add your handling code here:
         String nama=jTextFieldNama.getText();
@@ -464,8 +474,8 @@ ResultSet RsProduk=null;
                 Connection conn=(Connection)koneksi.koneksiDB();
                 Statement stt=conn.createStatement();
                 stt.executeUpdate("UPDATE dokter SET nik='"+nik+"', nama='"+nama+"' , alamat='"+alamat+"', telepon='"+telepon+"', tglahir='"+tl+"', goldar='"+darah+"', gender='"+jk+"', spesialisasi='"+spesialisasi+"' WHERE nama='"+nama+"'");
-                BersihData();
-                tampilData();
+                clearData();
+                showData();
                 SetEditOff();
                 JOptionPane.showMessageDialog(this,"Data berhasil disimpan","Success",JOptionPane.INFORMATION_MESSAGE);
             } catch(SQLException e){
@@ -478,13 +488,16 @@ ResultSet RsProduk=null;
         // TODO add your handling code here:
     }//GEN-LAST:event_jComboBoxJKActionPerformed
 
-     //menangani proses saat tabel produk di klik
+    //handling process when click data on pasien table
+    //menangani proses saat tabel pasien di klik
     private void TabelDokterMouseClicked(java.awt.event.MouseEvent evt) {                                         
         // TODO add your handling code here:
         seteditOn();
-        tblKeForm();
+        tableToForm();
     }       
     
+    //process for saving data
+    //proses untuk menyimpan data
     private void buttonSaveActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonSaveActionPerformed
         // TODO add your handling code here:
         String nama=jTextFieldNama.getText();
@@ -523,8 +536,8 @@ ResultSet RsProduk=null;
                 Statement stt=conn.createStatement();
                 stt.executeUpdate("INSERT INTO dokter(nik, nama, alamat, telepon, tglahir, goldar, gender, spesialisasi)"+
                     "VALUES('"+nik+"','"+nama+"','"+alamat+"','"+telepon+"','"+tl+"','"+darah+"','"+jk+"','"+spesialisasi+"')");
-                BersihData();
-                tampilData();
+                clearData();
+                showData();
                 SetEditOff();
                 JOptionPane.showMessageDialog(this,"Data berhasil disimpan","Success",JOptionPane.INFORMATION_MESSAGE);
             } catch(SQLException e){